// src/lib/types/fleet.ts
export interface CreateMaintenanceDto {
  vehicleId: string
  type?: string
  description?: string
  maintenanceDate: string
  costPkr: number
  shopOrPerson?: string
  odometerAtMaintenanceKm?: number
  nextServiceOdometerKm?: number
  nextServiceDueDate?: string
  maintenanceBy?: string
  status?: string
}